// QuizClient/src/components/types/api/conds.tsx
// ZOO実験条件について型指定

// この部分はデータベースに結果を登録していくようになったら排除される
export type LogStr = {
  meas_start_00: string;
  mount_start_00: string;
  mount_end_00: string;
  center_start_00: string;
  center_end_00: string;
  raster_start_00: string;
  raster_end_00: string;
  ds_start_00: string;
  ds_end_00: string;
  meas_end_00: string;
};

export type Conds = {
  _id: string;
  root_dir: string;
  o_index: number;
  p_index: number;
  mode: string;
  puckid: string;
  pinid: number;
  sample_name: string;
  wavelength: number;
  raster_vbeam: number;
  raster_hbeam: number;
  att_raster: number;
  hebi_att: number;
  exp_raster: number;
  dist_raster: number;
  loopsize: number;
  score_min: number;
  score_max: number;
  maxhits: number;
  total_osc: number;
  osc_width: number;
  ds_vbeam: number;
  ds_hbeam: number;
  exp_ds: number;
  dist_ds: number;
  dose_ds: number;
  offset_angle: number;
  reduced_fact: number;
  ntimes: number;
  meas_name: string;
  cry_min_size_um: number;
  cry_max_size_um: number;
  hel_full_osc: number;
  hel_part_osc: number;
  raster_roi: number;
  ln2_flag: number;
  cover_scan_flag: number;
  zoomcap_flag: number;
  warm_time: number;
  isSkip: number;
  isMount: number;
  isLoopCenter: number;
  isRaster: number;
  isDS: number;
  isDone: number;
  scan_height: number;
  scan_width: number;
  n_mount: number;
  nds_multi: number;
  nds_helical: number;
  nds_helpart: number;
  t_meas_start: LogStr;
  t_mount_end: string;
  t_cent_start: number;
  t_cent_end: number;
  t_raster_start: number;
  t_raster_end: number;
  t_ds_start: number;
  t_ds_end: number;
  t_dismount_start: number;
  t_dismount_end: number;
  data_index: number;
  n_mount_fails: number;
  log_mount: string;
  hel_cry_size: number;
  flux: number;
  phs_per_deg: number;
};